requirements.txt
opencv-python>=4.8.1
opencv-contrib-python>=4.8.1
numpy>=1.24.3
matplotlib>=3.7.2
pandas>=2.0.3
reportlab>=4.0.4
Pillow>=10.0.0
scipy>=1.11.1
seaborn>=0.12.2
configparser>=5.3.0
